Get rid of "su -c" for non-root users
Ideally, we want to give the opportunity to a trusted user to `make`
Archipelago as root. This will allow him to `make` the kernel components
as well as install them.

Since in my environment there is no `sudo`, the most obvious candidate
for this job is `su -c`. HOWEVER! `su -c`, due to a security patch for
TTY hijacking, behaves very different to `sudo`, in the sense that the
command running as root is detached from the shell/script that has been
invoked from. This means that, if the users wishes to end the script
during its execution, it will simply kill `su` and not the command.
Even worse, the command will probably print its output to the stdout of
the shell/script.

Until I find a fix for this or understand the situation better, the
non-root user must become root and then run this script.