Use extra evals to properly create logs
The evals (and they way they're used in this script) should generally be
avoided, but we can get away with it in our situation (we use them for
evaluation only). The reason we inserted more evals in this commit is
because there was an extra level of implicit argument dereferencing
which the first eval could not reach.